src/api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.20
pydantic==2.12.5
python-dotenv==1.0.1
dnspython==2.8.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
kubernetes==35.0.0
pydantic-settings==2.13.1
typing-extensions==4.15.0
docker
python-json-logger
PyYAML==6.0.3
nats-py==2.14.0
redis==7.4.0
requests==2.33.0
aiohttp==3.13.3
httpx==0.27.2
beautifulsoup4==4.14.3
Pillow>=10.0.0
tldextract==5.3.1
python-whois==0.9.5
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
APScheduler==3.10.4
croniter==2.0.1
pytz==2024.1